Correct memmove

+++ ft_memmove.c
@@ -6,7 +6,7 @@
 /*   By: tcharuel <tcharuel@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/11/06 18:25:05 by tcharuel          #+#    #+#             */
-/*   Updated: 2023/11/07 14:53:04 by tcharuel         ###   ########.fr       */
+/*   Updated: 2023/11/07 15:56:45 by tcharuel         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,7 +16,7 @@ void	*ft_memmove(void *dest, const void *src, size_t n)
 {
 	unsigned char	*buffer_src;
 	unsigned char	*buffer_dest;
-	size_t			i;
+	int				i;
 
 	if (dest == src || n == 0)
 		return (dest);
